Review of fixes for comment to last HW
