+++ GBShop/GBShop/App/Entity/CatalogData/CatalogDataResult.swift
@@ -9,11 +9,13 @@ import Foundation
 
 struct CatalogDataResult: Codable {
     
-    let pageNumber: Int
-    let products: [Product]
-    
-//    enum CodingKeys: String, CodingKey {
-//        case psgeNumber = "page_number"
-//        case products
-//    }
+    let productId: Int
+    let productName: String
+    let price: Int
+
+    enum CodingKeys: String, CodingKey {
+        case productId = "id_product"
+        case productName = "product_name"
+        case price = "price"
+    }
 }
