--- GBShop/GBShop/App/Entity/CatalogData/Product.swift
@@ -1,20 +1,0 @@
-//
-//  Product.swift
-//  GBShop
-//
-//  Created by Мария Коханчик on 17.08.2021.
-//
-
-import Foundation
-
-struct Product: Codable {
-    let id: Int
-    let productName: String
-    let price: Int
-    
-    enum CodingKeys: String, CodingKey {
-        case id = "id_product"
-        case productName = "product_name"
-        case price = "price"
-    }
-}
